implement wavegen support

+++ rtml_c_version/mlp.c
@@ -10,7 +10,7 @@ void forward_pass (struct mlp* m) {
 	while (current_layer=current_layer->next) {
 		// matrix-vector multiply
 		for (int i=0;i<current_layer->neurons;i++) {
-			float sum=0.f;
+			double sum=0.f;
 			for (int j=0;j<current_layer->prev->neurons;j++)
 				sum+=current_layer->prev->outputs[j] * current_layer->weights[i*current_layer->prev->neurons+j];
 			current_layer->outputs[i]=sum+current_layer->biases[i];
@@ -19,7 +19,7 @@ void forward_pass (struct mlp* m) {
 }
 #endif
 
-void backward_pass (struct mlp *m,float *y) {
+void backward_pass (struct mlp *m,double *y) {
 	// skip to last layer
 	struct layer *current_layer=&(m->layers[m->layerc-1]);
 
@@ -30,7 +30,7 @@ void backward_pass (struct mlp *m,float *y) {
 	while (current_layer->prev) {
 
 		for (int i=0;i<current_layer->neurons;i++) {
-			float sum=0.f;
+			double sum=0.f;
 			for (int j=0;j<current_layer->next->neurons;j++)
 				sum+=current_layer->next->deltas[j]*current_layer->next->weights[j];
 			current_layer->deltas[i]=current_layer->outputs[i]*sum;
@@ -42,11 +42,11 @@ void backward_pass (struct mlp *m,float *y) {
 
 void update_weights (struct mlp *m) {
 	struct layer *current_layer = &(m->layers[1]);
-	float alpha = m->alpha;
+	double alpha = m->alpha;
 
 	while (current_layer) {
 		for (int i=0;i<current_layer->neurons;i++) {
-			float sum=0.f;
+			double sum=0.f;
 			for (int j=0;j<current_layer->prev->neurons;j++)
 				current_layer->weights[i*current_layer->prev->neurons+j] -=
 					alpha * current_layer->deltas[i] * current_layer->prev->outputs[j];
@@ -58,13 +58,13 @@ void update_weights (struct mlp *m) {
 }
 
 #if 0
-float mac(float prev_outputs[1024],float current_weights[1024],int i,int prev_neurons) {
+double mac(double prev_outputs[1024],double current_weights[1024],int i,int prev_neurons) {
 	//#pragma HLS INTERFACE ap_memory port=prev_outputs
 	//#pragma HLS INTERFACE ap_memory port=current_weights
 	#pragma HLS array_partition variable=prev_outputs cyclic factor=32
 	#pragma HLS array_partition variable=current_weights cyclic factor=32
 
-	float sum=0.f,prods[32],sums[32];
+	double sum=0.f,prods[32],sums[32];
 	#pragma HLS array_partition variable=prods complete
 	#pragma HLS array_partition variable=sums complete
 
@@ -96,7 +96,7 @@ void forward_pass (struct mlp *m) {
 	// start with second layer, since the first layer is the input layer and there's
 	// nothing to do there
 	struct layer *current_layer=&(m->layers[0]);
-	float sum;
+	double sum;
 
 	while (current_layer=current_layer->next) {
 		// matrix-vector multiply
